temporal_analysis/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0
tqdm>=4.62.0
PyYAML>=6.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
fvcore>=0.1.5
timm>=0.6.0
einops>=0.4.0
